refactor. 잘못 분류된 util 함수 제거
