--- src/refactoring/utils/getQuantityDiscount.ts
@@ -1,13 +1,0 @@
-import { CartItem } from '../../types';
-
-export const getQuantityDiscount = (item: CartItem) => {
-  const { discounts } = item.product;
-  const { quantity } = item;
-  let appliedDiscount = 0;
-  for (const discount of discounts) {
-    if (quantity >= discount.quantity) {
-      appliedDiscount = Math.max(appliedDiscount, discount.rate);
-    }
-  }
-  return appliedDiscount;
-};
